+++ core/src/parquet_source/ver0_1_0.rs
@@ -4,12 +4,11 @@ use crate::deserialize::{Address, BigUnsigned, BloomFilterBytes, Bytes, Bytes32,
 use crate::types::{Block, Log, Transaction};
 use arrayvec::ArrayVec;
 use arrow2::array::{self, BooleanArray, UInt32Array, UInt64Array};
-use arrow2::compute::cast::cast;
 use arrow2::compute::concatenate::concatenate;
 use arrow2::datatypes::{DataType, Field};
 use std::collections::BTreeMap;
 
-type BinaryArray = array::BinaryArray<i64>;
+type BinaryArray = array::BinaryArray<i32>;
 
 pub struct Ver0_1_0;
